Change type rating

+++ client/src/components/EmployerCard/EmployerCard.js
@@ -5,7 +5,7 @@ import style from './EmployerCard.module.css';
 import {Link} from 'react-router-dom';
 
 const EmployerCard = ({employer}) => {
-  const rating = parseInt(employer.rating);
+  const rating = parseFloat(employer.rating);
   return (
     <>
       <Card
